helpdesk-agosto-2025/config.py
import os 
from dotenv import load_dotenv 

load_dotenv() 

class Config: 
    SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret-key-change-this") 
    DB_HOST = os.getenv("DB_HOST", "localhost") 
    DB_USER = os.getenv("DB_USER", "helpdesk_user") 
    DB_PASSWORD = os.getenv("DB_PASSWORD", "helpdesk_password") 
    DB_NAME = os.getenv("DB_NAME", "helpdesk_db") 