helpdesk-agosto-2025/requirements.txt
Flask 
pymysql 
python-dotenv 
Werkzeug 